module: move test reporter loading
Move the logic for handling --test-reporter out of the
general module loader and into the test_runner subsystem.
