+++ lib/internal/modules/run_main.js
@@ -2,10 +2,11 @@
 
 const {
   ObjectCreate,
+  StringPrototypeEndsWith,
 } = primordials;
+
 const { getOptionValue } = require('internal/options');
 const path = require('path');
-const { shouldUseESMLoader } = require('internal/modules/utils');
 
 function resolveMainPath(main) {
   // Note extension resolution for the main entry point can be deprecated in a
@@ -21,6 +22,29 @@ function resolveMainPath(main) {
     mainPath = toRealPath(mainPath);
 
   return mainPath;
+}
+
+function shouldUseESMLoader(mainPath) {
+  /**
+   * @type {string[]} userLoaders A list of custom loaders registered by the user
+   * (or an empty list when none have been registered).
+   */
+  const userLoaders = getOptionValue('--experimental-loader');
+  /**
+   * @type {string[]} userImports A list of preloaded modules registered by the user
+   * (or an empty list when none have been registered).
+   */
+  const userImports = getOptionValue('--import');
+  if (userLoaders.length > 0 || userImports.length > 0)
+    return true;
+  const { readPackageScope } = require('internal/modules/cjs/loader');
+  // Determine the module format of the main
+  if (mainPath && StringPrototypeEndsWith(mainPath, '.mjs'))
+    return true;
+  if (!mainPath || StringPrototypeEndsWith(mainPath, '.cjs'))
+    return false;
+  const pkg = readPackageScope(mainPath);
+  return pkg && pkg.data.type === 'module';
 }
 
 function runMainESM(mainPath) {
