+++ lib/internal/test_runner/utils.js
@@ -7,11 +7,10 @@ const {
   RegExpPrototypeExec,
   SafeMap,
 } = primordials;
-const { basename } = require('path');
+const { basename, isAbsolute } = require('path');
 const { createWriteStream } = require('fs');
 const { createDeferredPromise } = require('internal/util');
 const { getOptionValue } = require('internal/options');
-const { requireOrImport } = require('internal/modules/utils');
 
 const {
   codes: {
@@ -103,7 +102,13 @@ const kDefaultDestination = 'stdout';
 async function getReportersMap(reporters, destinations) {
   return SafePromiseAllReturnArrayLike(reporters, async (name, i) => {
     const destination = kBuiltinDestinations.get(destinations[i]) ?? createWriteStream(destinations[i]);
-    let reporter = await requireOrImport(kBuiltinReporters.get(name) ?? name);
+
+    // Load the test reporter passed to --test-reporter
+    let reporterPath = kBuiltinReporters.get(name) ?? name;
+    const { esmLoader } = require('internal/process/esm_loader');
+    const { pathToFileURL } = require('internal/url');
+    const file = isAbsolute(reporterPath) ? pathToFileURL(reporterPath).href : reporterPath;
+    let reporter = await esmLoader.import(file, undefined, { __proto__: null });
 
     if (reporter?.default) {
       reporter = reporter.default;
